Crate roux

roux.rs

This crate provides simple access to the Reddit API.

Using OAuth

To create an OAuth client with the reddit API, use the Reddit class.

use roux::Reddit;

let client = Reddit::new("USER_AGENT", "CLIENT_ID", "CLIENT_SECRET")
    .username("USERNAME")
    .password("PASSWORD")
    .login();

let me = client.unwrap();

It is important that you pick a good user agent. The ideal format is platform:program:version (by /u/yourname), e.g. macos:roux:v0.3.0 (by /u/beanpup_py).

This will authticate you as the user given in the username function.

Usage

Using the OAuth client, you can:

Submit A Text Post

use roux::Reddit;

let client = Reddit::new("USER_AGENT", "CLIENT_ID", "CLIENT_SECRET")
    .username("USERNAME")
    .password("PASSWORD")
    .login();

let me = client.unwrap();

me.submit_text("TEXT_TITLE", "TEXT_BODY", "SUBREDDIT");

Submit A Link Post

use roux::Reddit;

let client = Reddit::new("USER_AGENT", "CLIENT_ID", "CLIENT_SECRET")
    .username("USERNAME")
    .password("PASSWORD")
    .login();

let me = client.unwrap();

me.submit_link("LINK_TITLE", "LINK", "SUBREDDIT");

Re-exports

pub use subreddit::Subreddit;

pub use user::User;

Modules

subreddit
Subreddit module.

user
User module.

Structs

Reddit
Client to use OAuth with Reddit.